feat : 엔티티 상에 double -> float

+++ src/main/java/com/flexrate/flexrate_back/loan/domain/LoanApplication.java
@@ -40,7 +40,7 @@ public class LoanApplication {
     private LocalDateTime executedAt;
     private int totalAmount;
     private int remainAmount;
-    private double rate;
+    private float rate;
 
     @OneToMany(mappedBy = "application")
     private List<LoanTransaction> loanTransactions;
